backend/requirements.txt
flask-jwt-extended
gql
